+++ src/App.js
@@ -3,12 +3,12 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import "./css/common_styles.css";
 import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 import { slide as Menu } from 'react-burger-menu';
-import Home from './components/home.component';
-import Create from './components/create.component';
-import Edit from './components/edit.component';
-import Index from './components/index.component';
+import Home from './components/home/home.component';
+import Create from './components/create/create.component';
+import Edit from './components/edit/edit.component';
+import Index from './components/index/index.component';
 import Search from './components/search/search.component';
-import Timeline from './components/timeline.component';
+import Timeline from './components/timeline/timeline.component';
 import './css/tabs.css';
 import './css/mobile.css';
 // import ReactLoading from 'react-loading';
