+++ src/components/timeline/timeline.component.spec.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Timeline from './timeline.component.js'
+import { mount, shallow } from 'enzyme';
+
+describe ('<Timeline />', () => {
+    it (`should contain the Timeline component`, () => {
+        const component = shallow(<Timeline />);
+        expect((component).exists()).toBe(true);
+    });
+});
